Unify remote menu and start entry items for desktop (#185908)
* Unify remote menu and start entry items for desktop

* Clean up code

* Move remoteStartEntry to be web workbench contrib
